+++ feature/mypage/src/main/java/com/poptato/mypage/MyPageScreen.kt
@@ -1,0 +1,112 @@
+package com.poptato.mypage
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.poptato.design_system.Gray00
+import com.poptato.design_system.Gray100
+import com.poptato.design_system.Gray40
+import com.poptato.design_system.PoptatoTypo
+import com.poptato.design_system.R
+import com.poptato.design_system.SettingExampleMail
+import com.poptato.design_system.SettingExampleName
+
+@Composable
+fun MyPageScreen() {
+
+    MyPageContent()
+}
+
+@Composable
+fun MyPageContent() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Gray100)
+    ) {
+
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 16.dp)
+                .padding(end = 16.dp)
+        ) {
+            Icon(
+                painter = painterResource(id = R.drawable.ic_setting),
+                contentDescription = "",
+                tint = Color.Unspecified,
+                modifier = Modifier
+                    .size(24.dp)
+                    .align(Alignment.TopEnd)
+            )
+        }
+
+        MyData()
+
+        Image(
+            painter = painterResource(id = R.drawable.ic_setting_bg),
+            contentDescription = "ic_setting_bg",
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.Crop
+        )
+
+    }
+}
+
+@Composable
+fun MyData() {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(start = 16.dp)
+    ) {
+        Row {
+            Image(
+                painter = painterResource(id = R.drawable.ic_person),
+                contentDescription = "img_temp_person",
+                modifier = Modifier.size(64.dp)
+            )
+
+            Column {
+                Text(
+                    text = SettingExampleName,
+                    color = Gray00,
+                    style = PoptatoTypo.lgSemiBold,
+                    modifier = Modifier
+                        .offset(x = 12.dp, y = 8.dp)
+                )
+
+                Text(
+                    text = SettingExampleMail,
+                    color = Gray40,
+                    style = PoptatoTypo.smRegular,
+                    modifier = Modifier
+                        .offset(x = 12.dp, y = 10.dp)
+                )
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true, showSystemUi = true)
+@Composable
+fun PreviewSetting() {
+    MyPageContent()
+}
